Restructured pytest folder
